refactor: restructured pages folder structure
