+++ src/App.js
@@ -8,13 +8,15 @@ import {
   Playlists,
   SinglePlaylist,
   History,
+  WatchLater,
 } from './pages';
 import { Routes, Route } from 'react-router-dom';
-import { VideoCard } from './components';
+import { Header, VideoCard } from './components';
 
 function App() {
   return (
     <div className="app">
+      <Header />
       <Routes>
         <Route path="/" element={<Homepage />} />
         <Route path="/auth" element={<AuthPage />} />
@@ -25,6 +27,7 @@ function App() {
         <Route path="/playlists" element={<Playlists />} />
         <Route path="/playlist/:playlistName" element={<SinglePlaylist />} />
         <Route path="/history" element={<History />} />
+        <Route path="/watch-later" element={<WatchLater />} />
       </Routes>
     </div>
   );
