--- src/pages/SinglePlaylist.js
@@ -1,23 +1,0 @@
-import { Fragment } from 'react';
-import './SinglePlaylist.css';
-import { Header, VideoCard } from '../components';
-
-const SinglePlaylist = props => {
-  return (
-    <Fragment>
-      <Header />
-      <main className="main-single-playlist">
-        <div className="flex space-between align-center">
-          <h1>My Playlist</h1>
-          <button className="btn error rounded-edge">Delete Playlist</button>
-        </div>
-        <div className="videos-container">
-          {Array.from({ length: 4 }).map(el => (
-            <VideoCard dismissBtn={true} />
-          ))}
-        </div>
-      </main>
-    </Fragment>
-  );
-};
-export { SinglePlaylist };
