+++ src/pages/watch-later/WatchLater.css
@@ -1,0 +1,5 @@
+.main-watch-later {
+  max-width: 65rem;
+  margin: auto;
+  padding: 1rem;
+}
